Move dataclass
